# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Error response",
        "title": "Block Storage API",
        "contact": {
            "name": "Ciré LY",
            "url": "https://github.com/cire-ly/block-storage-api",
            "email": "cire.ly@nexonode.tech"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "host": "163.172.144.70:8080",
    "basePath": "/",
    "paths": {
        "/api/v1/volumes": {
            "get": {
                "description": "Returns all volumes managed by the storage backend.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "volumes"
                ],
                "summary": "List all volumes",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/volume.listResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/volume.errResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Creates a new block storage volume. The volume transitions PENDING → CREATING → AVAILABLE asynchronously.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "volumes"
                ],
                "summary": "Create a volume",
                "parameters": [
                    {
                        "description": "Volume creation parameters",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/volume.createVolumeRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/volume.volumeResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/volume.errResponse"
                        }
                    },
                    "409": {
                        "description": "Volume already exists",
                        "schema": {
                            "$ref": "#/definitions/volume.errResponse"
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "schema": {
                            "$ref": "#/definitions/volume.errResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/volumes/{name}": {
            "get": {
                "description": "Returns a single volume by name.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "volumes"
                ],
                "summary": "Get a volume",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Volume name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/volume.volumeResponse"
                        }
                    },
                    "404": {
                        "description": "Volume not found",
                        "schema": {
                            "$ref": "#/definitions/volume.errResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/volume.errResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Deletes an available volume. Transition: AVAILABLE → DELETING → DELETED (async). Returns 409 if attached.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "volumes"
                ],
                "summary": "Delete a volume",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Volume name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted"
                    },
                    "404": {
                        "description": "Volume not found",
                        "schema": {
                            "$ref": "#/definitions/volume.errResponse"
                        }
                    },
                    "409": {
                        "description": "Invalid FSM transition (e.g. volume is attached)",
                        "schema": {
                            "$ref": "#/definitions/volume.errResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/volume.errResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/volumes/{name}/attach": {
            "put": {
                "description": "Attaches an available volume to a compute node. Transition: AVAILABLE → ATTACHING → ATTACHED (async).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "volumes"
                ],
                "summary": "Attach a volume to a node",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Volume name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Node to attach to",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/volume.attachVolumeRequest"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "$ref": "#/definitions/volume.volumeResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/volume.errResponse"
                        }
                    },
                    "404": {
                        "description": "Volume not found",
                        "schema": {
                            "$ref": "#/definitions/volume.errResponse"
                        }
                    },
                    "409": {
                        "description": "Invalid FSM transition",
                        "schema": {
                            "$ref": "#/definitions/volume.errResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/volume.errResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/volumes/{name}/detach": {
            "put": {
                "description": "Detaches an attached volume. Transition: ATTACHED → DETACHING → AVAILABLE (async).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "volumes"
                ],
                "summary": "Detach a volume from its node",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Volume name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "$ref": "#/definitions/volume.volumeResponse"
                        }
                    },
                    "404": {
                        "description": "Volume not found",
                        "schema": {
                            "$ref": "#/definitions/volume.errResponse"
                        }
                    },
                    "409": {
                        "description": "Invalid FSM transition",
                        "schema": {
                            "$ref": "#/definitions/volume.errResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/volume.errResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/volumes/{name}/reset": {
            "post": {
                "description": "Resets a volume from the terminal error state back to pending, clearing the retry counter.\nReturns 409 if the volume is not in the error state.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "volumes"
                ],
                "summary": "Reset a volume from error state",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Volume name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/volume.volumeResponse"
                        }
                    },
                    "404": {
                        "description": "Volume not found",
                        "schema": {
                            "$ref": "#/definitions/volume.errResponse"
                        }
                    },
                    "409": {
                        "description": "Volume not in error state",
                        "schema": {
                            "$ref": "#/definitions/volume.errResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/volume.errResponse"
                        }
                    }
                }
            }
        },
        "/healthz": {
            "get": {
                "description": "Returns the health status of the API and storage backend.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "system"
                ],
                "summary": "Health check",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/volume.healthResponse"
                        }
                    },
                    "503": {
                        "description": "Backend degraded",
                        "schema": {
                            "$ref": "#/definitions/volume.healthResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "volume.attachVolumeRequest": {
            "description": "Attach volume request",
            "type": "object",
            "properties": {
                "node_id": {
                    "type": "string",
                    "example": "node-paris-01"
                }
            }
        },
        "volume.createVolumeRequest": {
            "description": "Create volume request",
            "type": "object",
            "properties": {
                "name": {
                    "type": "string",
                    "example": "vol-01"
                },
                "size_mb": {
                    "type": "integer",
                    "example": 1024
                }
            }
        },
        "volume.errResponse": {
            "description": "Error response",
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "volume not found"
                }
            }
        },
        "volume.healthResponse": {
            "description": "Health check response",
            "type": "object",
            "properties": {
                "backend": {
                    "type": "string",
                    "example": "mock"
                },
                "status": {
                    "type": "string",
                    "example": "ok"
                },
                "version": {
                    "type": "string",
                    "example": "1.0.0"
                }
            }
        },
        "volume.listResponse": {
            "description": "List of volumes",
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer",
                    "example": 1
                },
                "volumes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/volume.volumeResponse"
                    }
                }
            }
        },
        "volume.volumeResponse": {
            "description": "Volume resource",
            "type": "object",
            "properties": {
                "backend": {
                    "type": "string",
                    "example": "mock"
                },
                "created_at": {
                    "type": "string",
                    "example": "2024-01-01T00:00:00Z"
                },
                "id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "name": {
                    "type": "string",
                    "example": "vol-01"
                },
                "node_id": {
                    "type": "string",
                    "example": "node-paris-01"
                },
                "size_mb": {
                    "type": "integer",
                    "example": 1024
                },
                "state": {
                    "type": "string",
                    "example": "available"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2024-01-01T00:00:00Z"
                }
            }
        }
    },
    "tags": [
        {
            "description": "Block storage volume lifecycle operations",
            "name": "volumes"
        },
        {
            "description": "Health and observability endpoints",
            "name": "system"
        }
    ]
}